+++ generators/webtranslateit/webtranslateit_generator.rb
@@ -12,7 +12,7 @@ class WebtranslateitGenerator < Rails::Generator::Base
     end
     record do |m|
       if options[:api_key]
-        project_details = YAML.load fetch_project_information(options[:api_key])
+        project_details = YAML.load WebTranslateIt::Project.fetch_info(options[:api_key])
         m.template 'translation.yml', 'config/translation.yml',
           :assigns => { :api_key => options[:api_key], :project => project_details["project"] }
         m.append_to 'Rakefile', "require 'web_translate_it/tasks' rescue LoadError"
@@ -22,23 +22,5 @@ class WebtranslateitGenerator < Rails::Generator::Base
 
   def api_key_configured?
     File.exists?('config/translations.yml')
-  end
-  
-  def fetch_project_information(api_key)
-    WebTranslateIt::Util.http_connection do |http|
-      request  = Net::HTTP::Get.new("/api/projects/#{api_key}.yaml")
-      response = http.request(request)
-      if response.code.to_i >= 400 and response.code.to_i < 500
-        puts "We had a problem connecting to Web Translate It with this API key."
-        puts "Make sure it is correct."
-        exit
-      elsif response.code.to_i >= 500
-        puts "Web Translate It is temporarily unavailable. Please try again shortly."
-        exit
-      else
-        return response.body
-      end
-    end
-    
   end
 end