+++ lib/web_translate_it.rb
@@ -3,6 +3,7 @@ require File.join(File.dirname(__FILE__), 'web_translate_it', 'configuration')
 require File.join(File.dirname(__FILE__), 'web_translate_it', 'translation_file')
 require File.join(File.dirname(__FILE__), 'web_translate_it', 'auto_fetch')
 require File.join(File.dirname(__FILE__), 'web_translate_it', 'command_line')
+require File.join(File.dirname(__FILE__), 'web_translate_it', 'project')
 
 module WebTranslateIt
   def self.fetch_translations