+++ lib/web_translate_it/command_line.rb
@@ -4,6 +4,7 @@ module WebTranslateIt
     OPTIONS = <<-OPTION
 pull            Pull language file(s) from Web Translate It.
 push            Push language file(s) to Web Translate It.
+autoconf        Configure your project to sync with Web Translate It.
 
 OPTIONAL PARAMETERS:
 --------------------
@@ -24,6 +25,8 @@ OPTION
         pull
       when 'push'
         push
+      when 'autoconf'
+        autoconf
       when '-v', '--version'
         show_version
       when '-h', '--help'
@@ -59,6 +62,23 @@ OPTION
           file.upload(locale)
         end
       end
+    end
+    
+    def self.autoconf
+      puts "We will attempt to configure your project automagically"
+      puts "Please enter your project API Key:"
+      api_key = STDIN.gets.strip
+      if api_key == ""
+        puts "You must enter your project API key provided by Web Translate It"
+        exit
+      end
+      puts "We will now create a `config/translation.yml` file in the current directory."
+      puts "Enter below another path if you want to change, or leave it blank if the defaut path is okay."
+      path = STDIN.gets.strip
+      path = "config/translation.yml" if path == ""
+      File.open(path, 'w'){ |file| file << generate_configuration(api_key) }
+      puts "Done! You can now use `wti` to push and pull your language files."
+      puts "Check `wti --help` for more information."
     end
     
     def self.show_options
@@ -125,6 +145,33 @@ OPTION
         locales = [ARGV[index+1]]
       end
       return locales
+    end
+    
+    def self.generate_configuration(api_key)
+      project_info = YAML.load WebTranslateIt::Project.fetch_info(api_key)
+      project = project_info['project']
+      file = <<-FILE
+api_key: #{api_key}
+
+# The locales not to sync with Web Translate It.
+# Pass an array of string, or an array of symbols, a string or a symbol.
+# eg. [:en, :fr] or just 'en'
+ignore_locales: '#{project["source_locale"]["code"]}'
+
+# A list of files to translate
+# You can name your language files as you want, as long as the locale name match the
+# locale name you set in Web Translate It, and that the different language files names are
+# differenciated by their locale name.
+# For example, if you set to translate a project in en_US in WTI, you should use the locale en_US in your app
+#
+files:
+FILE
+      project["project_files"].each do |project_file|
+        if project_file["master"]
+          file << "  #{project_file["id"]}: config/locales/" + project_file["name"].gsub(project["source_locale"]["code"], "[locale]") + "\n"
+        end
+      end
+      return file
     end
   end
 end